Comment and modularize code.
